// src/types/schema.ts
export type ID1 = string;
export type SchemaKey1 = "Person";
/**
 * An ORCID (orcid.org) identifier for an individual.
 */
export type AnORCIDIdentifier = string;
/**
 * Use the format: familyname, given names ...
 */
export type Name = string;
export type Email = string;
export type URL = string;
/**
 * An enumeration of roles
 */
export type RoleType =
  | "dcite:Author"
  | "dcite:Conceptualization"
  | "dcite:ContactPerson"
  | "dcite:DataCollector"
  | "dcite:DataCurator"
  | "dcite:DataManager"
  | "dcite:FormalAnalysis"
  | "dcite:FundingAcquisition"
  | "dcite:Investigation"
  | "dcite:Maintainer"
  | "dcite:Methodology"
  | "dcite:Producer"
  | "dcite:ProjectLeader"
  | "dcite:ProjectManager"
  | "dcite:ProjectMember"
  | "dcite:ProjectAdministration"
  | "dcite:Researcher"
  | "dcite:Resources"
  | "dcite:Software"
  | "dcite:Supervision"
  | "dcite:Validation"
  | "dcite:Visualization"
  | "dcite:Funder"
  | "dcite:Sponsor"
  | "dcite:StudyParticipant"
  | "dcite:Affiliation"
  | "dcite:EthicsApproval"
  | "dcite:Other";
/**
 * Role(s) of the contributor. Multiple roles can be selected.
 */
export type Role = RoleType[];
/**
 * A flag to indicate whether a contributor should be included when generating a citation for the item.
 */
export type IncludeContributorInCitation = boolean;
/**
 * Identifier associated with a sponsored or gift award.
 */
export type IdentifierForAnAward = string;
/**
 * Uniform resource identifier
 */
export type ID2 = string;
export type SchemaKey2 = "Affiliation";
/**
 * Use an ror.org identifier for institutions.
 */
export type ARorOrgIdentifier = string;
/**
 * Name of organization
 */
export type Name1 = string;
/**
 * An organization that this person is affiliated with.
 */
export type Affiliation = Affiliation1[];
/**
 * Uniform resource identifier
 */
export type ID3 = string;
export type SchemaKey3 = "Organization";
/**
 * Use an ror.org identifier for institutions.
 */
export type ARorOrgIdentifier1 = string;
export type Name2 = string;
export type Email1 = string;
export type URL1 = string;
/**
 * Role(s) of the contributor. Multiple roles can be selected.
 */
export type Role1 = RoleType[];
/**
 * A flag to indicate whether a contributor should be included when generating a citation for the item
 */
export type IncludeContributorInCitation1 = boolean;
/**
 * Identifier associated with a sponsored or gift award.
 */
export type IdentifierForAnAward1 = string;
/**
 * Uniform resource identifier
 */
export type ID4 = string;
export type SchemaKey4 = "ContactPoint";
/**
 * Email address of contact.
 */
export type Email2 = string;
/**
 * A Web page to find information on how to contact.
 */
export type URL2 = string;
/**
 * Contact for the organization
 */
export type OrganizationContactInformation = ContactPoint[];
/**
 * Uniform resource identifier
 */
export type ID5 = string;
export type SchemaKey5 = "Disorder";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier = string;
/**
 * The name of the item.
 */
export type Name3 = string;
/**
 * Dates of diagnosis
 */
export type DatesOfDiagnosis = (string | string)[];
/**
 * Uniform resource identifier
 */
export type ID6 = string;
export type SchemaKey6 = "Anatomy";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier1 = string;
/**
 * The name of the item.
 */
export type Name4 = string;
/**
 * Uniform resource identifier
 */
export type ID7 = string;
export type SchemaKey7 = "GenericType";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier2 = string;
/**
 * The name of the item.
 */
export type Name5 = string;
// ...
export type ID8 = string;
export type SchemaKey8 = "EthicsApproval";
/**
 * Approved Protocol identifier, often a number or alphanumeric string.
 */
export type ApprovedProtocolIdentifier = string;
// ...
export type ID10 = string;
export type SchemaKey10 = "Resource";
export type Identifier3 = string;
export type ATitleOfTheResource = string;
export type URLOfTheResource = string;
/**
 * Name of the repository in which the resource is housed.
 */
export type NameOfTheRepository = string;
/**
 * Indicates how the resource is related to the dataset. This relation should satisfy: dandiset <relation> resource.
 */
export type ResourceRelation =
  | "dcite:IsCitedBy"
  | "dcite:Cites"
  | "dcite:IsSupplementTo"
  | "dcite:IsSupplementedBy"
  | "dcite:IsContinuedBy"
  | "dcite:Continues"
  | "dcite:Describes"
  | "dcite:IsDescribedBy"
  | "dcite:HasMetadata"
  | "dcite:IsMetadataFor"
  | "dcite:HasVersion"
  | "dcite:IsVersionOf"
  | "dcite:IsNewVersionOf"
  | "dcite:IsPreviousVersionOf"
  | "dcite:IsPartOf"
  | "dcite:HasPart"
  | "dcite:IsReferencedBy"
  | "dcite:References"
  | "dcite:IsDocumentedBy"
  | "dcite:Documents"
  | "dcite:IsCompiledBy"
  | "dcite:Compiles"
  | "dcite:IsVariantFormOf"
  | "dcite:IsOriginalFormOf"
  | "dcite:IsIdenticalTo"
  | "dcite:IsReviewedBy"
  | "dcite:Reviews"
  | "dcite:IsDerivedFrom"
  | "dcite:IsSourceOf"
  | "dcite:IsRequiredBy"
  | "dcite:Requires"
  | "dcite:Obsoletes"
  | "dcite:IsObsoletedBy"
  | "dcite:IsPublishedIn";
// ...
export type ID12 = string;
export type SchemaKey12 = "Software";
/**
 * RRID of the software from scicrunch.org.
 */
export type ResearchResourceIdentifier = string;
export type Name6 = string;
export type Version = string;
/**
 * Web page for the software.
 */
export type URL4 = string;
// ...
export interface Person {
  id?: ID1;
  schemaKey: SchemaKey1;
  identifier?: AnORCIDIdentifier;
  name: Name;
  email?: Email;
  url?: URL;
  roleName?: Role;
  includeInCitation?: IncludeContributorInCitation;
  awardNumber?: IdentifierForAnAward;
  affiliation?: Affiliation;
  [k: string]: unknown;
}
export interface Affiliation1 {
  id?: ID2;
  schemaKey: SchemaKey2;
  identifier?: ARorOrgIdentifier;
  name: Name1;
  [k: string]: unknown;
}
export interface Organization {
  id?: ID3;
  schemaKey: SchemaKey3;
  identifier?: ARorOrgIdentifier1;
  name?: Name2;
  email?: Email1;
  url?: URL1;
  roleName?: Role1;
  includeInCitation?: IncludeContributorInCitation1;
  awardNumber?: IdentifierForAnAward1;
  contactPoint?: OrganizationContactInformation;
  [k: string]: unknown;
}
export interface ContactPoint {
  id?: ID4;
  schemaKey: SchemaKey4;
  email?: Email2;
  url?: URL2;
  [k: string]: unknown;
}
/**
 * Biolink, SNOMED, or other identifier for disorder studied
 */
export interface Disorder {
  id?: ID5;
  schemaKey: SchemaKey5;
  identifier?: Identifier;
  name?: Name3;
  dxdate?: DatesOfDiagnosis;
  [k: string]: unknown;
}
/**
 * UBERON or other identifier for anatomical part studied
 */
export interface Anatomy {
  id?: ID6;
  schemaKey: SchemaKey6;
  identifier?: Identifier1;
  name?: Name4;
  [k: string]: unknown;
}
/**
 * An object to capture any type for about
 */
export interface GenericType {
  id?: ID7;
  schemaKey: SchemaKey7;
  identifier?: Identifier2;
  name?: Name5;
  [k: string]: unknown;
}
/**
 * Information about ethics committee approval for project
 */
export interface EthicsApproval {
  id?: ID8;
  schemaKey: SchemaKey8;
  identifier: ApprovedProtocolIdentifier;
  contactPoint?: ContactPoint1;
  [k: string]: unknown;
}
/**
 * Information about the ethics approval committee.
 */
export interface ContactPoint1 {
  id?: ID4;
  schemaKey: SchemaKey4;
  email?: Email2;
  url?: URL2;
  [k: string]: unknown;
}
// ...
export interface Resource {
  id?: ID10;
  schemaKey: SchemaKey10;
  identifier?: Identifier3;
  name?: ATitleOfTheResource;
  url?: URLOfTheResource;
  repository?: NameOfTheRepository;
  relation: ResourceRelation;
  [k: string]: unknown;
}
/**
 * Information about the Project activity
 */
export interface Project {
  id?: ID11;
  schemaKey: SchemaKey11;
  identifier?: Identifier4;
  name: NameOfProject;
  description?: Description2;
  startDate?: StartDate;
  endDate?: EndDate;
  wasAssociatedWith?: WasAssociatedWith;
  used?: Used;
  [k: string]: unknown;
}
export interface Software {
  id?: ID12;
  schemaKey: SchemaKey12;
  identifier?: ResearchResourceIdentifier;
  name: Name6;
  version: Version;
  url?: URL4;
  [k: string]: unknown;
}